+++ MapboxNavigation/NextBannerView.swift
@@ -14,6 +14,11 @@ open class NextBannerView: UIView {
     weak var maneuverView: ManeuverView!
     weak var instructionLabel: NextInstructionLabel!
     weak var bottomSeparatorView: SeparatorView!
+    weak var instructionDelegate: VisualInstructionDelegate? {
+        didSet {
+            instructionLabel.instructionDelegate = instructionDelegate
+        }
+    }
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -39,6 +44,7 @@ open class NextBannerView: UIView {
         self.maneuverView = maneuverView
         
         let instructionLabel = NextInstructionLabel()
+        instructionLabel.instructionDelegate = instructionDelegate
         instructionLabel.shieldHeight = instructionLabel.font.pointSize
         instructionLabel.translatesAutoresizingMaskIntoConstraints = false
         addSubview(instructionLabel)
@@ -85,11 +91,13 @@ open class NextBannerView: UIView {
         bottomSeparatorView.heightAnchor.constraint(equalToConstant: 1 / UIScreen.main.scale).isActive = true
     }
     
-    public func update(for visualInstruction: VisualInstructionBanner, currentStepProgress: RouteStepProgress) {
+    public func update(for currentStepProgress: RouteStepProgress) {
         let distanceTraveled = currentStepProgress.distanceTraveled
-        guard let tertiaryInstruction = visualInstruction.tertiaryInstruction,
+        guard let visualInstruction = currentStepProgress.currentVisualInstruction,
+              let tertiaryInstruction = visualInstruction.tertiaryInstruction,
                   distanceTraveled <= visualInstruction.distanceAlongStep,
-                  !tertiaryInstruction.containsLaneIndications else {
+                  !tertiaryInstruction.containsLaneIndications
+            else {
                     hide()
                     return
         }